feat: TodoInput 컴포넌트 작성 (#2)
* Object, Array, Map, Set등은 reactive를 사용하여 변경 감지
* Object, String, Number 는 ref를 사용하여 변경 감지를 해야 한다.
* String 타입을 reactive로 선언하면 변경 감지를 제대로 하지 못한다.
